Add .go-version and .golangci.yml (#32)
* Add .go-version and .golangci.yml

Enable golangci-lint and specify the Go version in a .go-version file.

* Update Go version in Docker image

Update the Go version used to build the binary in the Docker image.

Use dependencies from the runtime image rather than copying them from
the build image. This was holdover from when we used to distribute
an image built `FROM scratch`.