automation - add changelog and release automation
Add a new Github workflow for releases that automates
the generation of a draft GH release with changelog notes.

Pull requests will be required to add new file to at
.changelog/<pr_number>.txt with a changelog message or
label with skip-changelog to bypass the check. See
CONTRIBUTING.md for more details.

This backfills the .changelog directory with entries
for changes since v0.14.0.